refacto algo displayed tag
