+++ src/components/specific/files/files-table/file-tags-cell/FileTagsCell.vue
@@ -1,6 +1,7 @@
 <template>
   <div class="file-tags-cell">
-    <template v-for="tag in tagsToDisplay" :key="tag.id">
+    {{ console.log("tagList", tagList) }}
+    <template v-for="tag in tagList.displayedArray" :key="tag.id">
       <div
         class="file-tags-cell__tag"
         :style="{
@@ -21,6 +22,7 @@
 <script>
 import { computed } from "vue";
 import { isEmpty } from "lodash";
+import { addSizeToTag, isTooWide, tagDispatcher } from "./tag-config";
 
 export default {
   props: {
@@ -30,68 +32,22 @@ export default {
     }
   },
   setup(props) {
-    const tagStyle = {
-      maxBoxWidth: 126 * 2,
-      maxTagWidth: 58,
-      fontSize: 10,
-      paddingX: 5,
-      gap: 3
-    };
-
-    const isTooWide = sumTagSize => tagStyle.maxBoxWidth < sumTagSize;
-
-    const tagDispatcher = (tagsWithSize, sumTagSize) => {
-      let tagList = {
-        displayedArray: [],
-        hiddedArray: []
-      };
-      for (const tag of tagsWithSize.reverse()) {
-        sumTagSize -= tag.size;
-
-        if (isTooWide(sumTagSize)) {
-          tagList.hiddedArray.push(tag);
-        } else {
-          tagList.displayedArray = tagsWithSize.slice(
-            0,
-            tagsWithSize.length - tagList.hiddedArray.length
-          );
-          break;
-        }
-      }
-      return tagList;
-    };
-
-    const tagsToDisplay = computed(() => {
+    const tagList = computed(() => {
       if (!isEmpty(props.file.tags)) {
-        const sizeOfStyle = tagStyle.paddingX * 2 + tagStyle.gap;
-
-        const tagsWithSize = props.file.tags.map(tag => {
-          const size = tag.name.length * (tagStyle.fontSize / 2);
-
-          return {
-            ...tag,
-            size:
-              size < tagStyle.maxTagWidth
-                ? size + sizeOfStyle
-                : tagStyle.maxTagWidth + sizeOfStyle
-          };
-        });
-
+        const tagsWithSize = addSizeToTag(props.file.tags);
         let sumTagSize = tagsWithSize.reduce((a, b) => a + b.size, 0);
 
         if (isTooWide(sumTagSize)) {
-          const res = tagDispatcher(tagsWithSize, sumTagSize);
-          console.log("res", res);
-          return props.file.tags;
+          return tagDispatcher(tagsWithSize, sumTagSize);
         } else {
-          return props.file.tags;
+          return { displayedArray: props.file.tags };
         }
       } else {
-        return props.file.tags;
+        return { displayedArray: [] };
       }
     });
     return {
-      tagsToDisplay,
+      tagList,
       console
     };
   }
